pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rest-extensions"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Xiang Wang", email = "ramwin@qq.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = ["django", "djangorestframework"]

[project.urls]
Documentation = "https://github.com/ramwin/django-rest-extensions#readme"
Issues = "https://github.com/ramwin/django-rest-extensions/issues"
Source = "https://github.com/ramwin/django-rest-extensions"

[tool.hatch.version]
path = "src/rest_extensions/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/rest_extensions tests}"

[tool.coverage.run]
source_pkgs = ["rest_extensions", "tests"]
branch = true
parallel = true
omit = [
  "src/rest_extensions/__about__.py",
]

[tool.coverage.paths]
rest_extensions = ["src/rest_extensions", "*/rest-extensions/src/rest_extensions"]
tests = ["tests", "*/rest-extensions/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
[tool.setuptools.package-data]
django_commands = ["src/rest_extensions/py.typed"]
